Rename the package and change 'nlist' to 'list'.
