+++ common/src/main/java/org/pierce/list/MatchType.java
@@ -1,4 +1,4 @@
-package org.pierce.nlist;
+package org.pierce.list;
 
 public enum MatchType {
     REGULAR_MATCHING,
