+++ common/src/main/java/org/pierce/list/NameListCheck.java
@@ -1,4 +1,4 @@
-package org.pierce.nlist;
+package org.pierce.list;
 
 public interface NameListCheck {
