+++ common/src/main/java/org/pierce/list/entity/NameEntity.java
@@ -1,4 +1,4 @@
-package org.pierce.nlist.entity;
+package org.pierce.list.entity;
 
 public class NameEntity {
     int seq;
