+++ common/src/main/java/org/pierce/list/gfw/Base64InputStream.java
@@ -1,4 +1,4 @@
-package org.pierce.nlist.gfw;
+package org.pierce.list.gfw;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
