+++ common/src/main/java/org/pierce/list/gfw/ByteArrayScanner.java.bak
@@ -1,4 +1,4 @@
-package org.pierce.nlist;
+package org.pierce.list;
 
 public class ByteArrayScanner {
     byte[] buffer;
