+++ common/src/main/java/org/pierce/list/gfw/NameListCheck.java.bak
@@ -1,7 +1,7 @@
-package org.pierce.nlist.gfw;
+package org.pierce.list.gfw;
 
 
-import org.pierce.nlist.ByteArrayScanner;
+import org.pierce.list.ByteArrayScanner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
