+++ common/src/main/java/org/pierce/list/imp/DefaultNameListCheck.java
@@ -1,8 +1,8 @@
-package org.pierce.nlist.imp;
+package org.pierce.list.imp;
 
 import org.apache.commons.validator.routines.InetAddressValidator;
-import org.pierce.nlist.Directive;
-import org.pierce.nlist.NameListCheck;
+import org.pierce.list.Directive;
+import org.pierce.list.NameListCheck;
 
 public class DefaultNameListCheck implements NameListCheck {
