+++ local-server/src/main/java/org/pierce/LocalServer.java
@@ -9,9 +9,9 @@ import org.pierce.imp.DefaultSelector;
 import org.pierce.impl.DefaultConnectionTypeCheck;
 import org.pierce.mybatis.entity.HostName2Address;
 import org.pierce.mybatis.mapper.HostName2AddressMapper;
-import org.pierce.nlist.imp.DataBaseNameListCheck;
-import org.pierce.nlist.imp.GFWNameListCheck;
-import org.pierce.nlist.imp.TextNameListCheck;
+import org.pierce.list.imp.DataBaseNameListCheck;
+import org.pierce.list.imp.GFWNameListCheck;
+import org.pierce.list.imp.TextNameListCheck;
 import org.pierce.socks.SocksServer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
