+++ local-server/src/main/java/org/pierce/httpproxy/HttpProxyServer.java
@@ -7,8 +7,8 @@ import io.netty.channel.socket.nio.NioServerSocketChannel;
 import org.pierce.JproxyProperties;
 import org.pierce.JproxyServer;
 import org.pierce.httpproxy.handler.HttpProxyServerInitializer;
-import org.pierce.nlist.NameListCheck;
-import org.pierce.nlist.imp.FixedReturnConnectListCheck;
+import org.pierce.list.NameListCheck;
+import org.pierce.list.imp.FixedReturnConnectListCheck;
 import org.pierce.socks.SocksServer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
