+++ local-server/src/main/java/org/pierce/impl/DefaultConnectionTypeCheck.java
@@ -5,8 +5,8 @@ import io.netty.channel.EventLoop;
 import io.netty.util.concurrent.*;
 import org.pierce.ConnectionTypeCheck;
 import org.pierce.entity.ConnectType;
-import org.pierce.nlist.Directive;
-import org.pierce.nlist.NameListCheck;
+import org.pierce.list.Directive;
+import org.pierce.list.NameListCheck;
 import org.pierce.session.SessionAttributes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
