+++ local-server/src/main/java/org/pierce/socks/SocksServer.java
@@ -6,9 +6,8 @@ import io.netty.channel.nio.NioIoHandler;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
 import org.pierce.JproxyProperties;
 import org.pierce.JproxyServer;
-import org.pierce.LocalServer;
-import org.pierce.nlist.NameListCheck;
-import org.pierce.nlist.imp.FixedReturnConnectListCheck;
+import org.pierce.list.NameListCheck;
+import org.pierce.list.imp.FixedReturnConnectListCheck;
 import org.pierce.socks.handler.SocksServerInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
