+++ local-server/src/main/java/org/pierce/socks/handler/SocksServerHandler.java
@@ -16,7 +16,7 @@ import org.pierce.entity.LocalLinkStatusEvent;
 import org.pierce.entity.LocalLinkStep;
 import org.pierce.handler.LinkOutHandler;
 import org.pierce.handler.LinkOutOverWebSocketHandler;
-import org.pierce.nlist.Directive;
+import org.pierce.list.Directive;
 import org.pierce.session.SessionAttributes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
