+++ remote-server/src/main/java/org/pierce/RemoteServer.java
@@ -3,8 +3,8 @@ package org.pierce;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.MultiThreadIoEventLoopGroup;
 import io.netty.channel.nio.NioIoHandler;
-import org.pierce.nlist.NameListCheck;
-import org.pierce.nlist.imp.TextNameListCheck;
+import org.pierce.list.NameListCheck;
+import org.pierce.list.imp.TextNameListCheck;
 import org.pierce.socks.RemoteSocks;
 import org.pierce.websocket.RemoteWebSocket;
 import org.slf4j.Logger;
