+++ remote-server/src/main/java/org/pierce/websocket/handler/WebSocketFrameHandler.java
@@ -11,7 +11,7 @@ import io.netty.util.concurrent.Promise;
 import org.pierce.RemoteServer;
 import org.pierce.UtilTools;
 import org.pierce.handler.LinkOutHandler;
-import org.pierce.nlist.Directive;
+import org.pierce.list.Directive;
 import org.pierce.session.SessionAttributes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
